feat: melhorias na regra de negócio

+++ src/main/java/com/example/obra/repository/AutorRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AutorRepository extends JpaRepository<AutorModel, Long> {
+public interface AutorRepository extends JpaRepository<AutorModel, Integer> {
 
     AutorModel findByCpf(String cpf);
